requirements.txt
streamlit
opencv-python
numpy
scikit-learn
streamlit-cropper
Pillow